refactor : 상수만 담는 클래스니까 constants로 고쳐라
