+++ src/main/java/com/codessquad/qna/web/utils/DateConstants.java
@@ -2,10 +2,10 @@ package com.codessquad.qna.web.utils;
 
 import java.time.format.DateTimeFormatter;
 
-public class DateUtil {
+public class DateConstants {
 
     public static final DateTimeFormatter MY_DATE_FORMATTER = DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm:ss");
 
-    private DateUtil() {
+    private DateConstants() {
     }
 }
